+++ HISTORY.rst
@@ -3,6 +3,23 @@
 Release History
 ---------------
 
+2.13.2 (2021-08-26)
++++++++++++++++++++
+
+**Improvements**
+
+- currency_parameters updated
+
+**Bug Fixes**
+
+- Bump socket timeout to 64s to correctly raise Betfair timeouts and prevent unwanted Socket timeouts
+- Split socket shutdown and close in socket.stop()
+- Missing timeouts added to request endpoints
+
+**Dependencies**
+
+- orjson upgraded to 3.6.3
+
 2.13.1 (2021-08-16)
 +++++++++++++++++++